Address issues from code review of LiveReload.
* Improve variable names
* Improve resiliency when conflicting options are given
* Refactor the LiveReloadReactor class into a separate file
* Correct typos
* Use alias_method for brevity
* Use __dir__ instead of File.dirname(__FILE__)
* Set and use default LiveReload port
