+++ test/test_commands_serve.rb
@@ -50,14 +50,13 @@ class TestCommandsServe < JekyllUnitTest
       @client = HTTPClient.new
       @client.connect_timeout = 5
       @standard_options = {
-        "port"            => 4000,
-        "host"            => "localhost",
-        "baseurl"         => "",
-        "detach"          => false,
-        "livereload"      => true,
-        "source"          => @temp_dir,
-        "destination"     => @destination,
-        "livereload_port" => Jekyll::Commands::Serve.singleton_class::LIVERELOAD_PORT,
+        "port"        => 4000,
+        "host"        => "localhost",
+        "baseurl"     => "",
+        "detach"      => false,
+        "livereload"  => true,
+        "source"      => @temp_dir,
+        "destination" => @destination,
       }
 
       site = instance_double(Jekyll::Site)
@@ -95,6 +94,7 @@ class TestCommandsServe < JekyllUnitTest
     end
 
     should "serve livereload.js over HTTP on the default LiveReload port" do
+      skip_if_windows "EventMachine support on Windows is limited"
       opts = serve(@standard_options)
       content = @client.get_content(
         "http://#{opts["host"]}:#{opts["livereload_port"]}/livereload.js"
@@ -103,6 +103,7 @@ class TestCommandsServe < JekyllUnitTest
     end
 
     should "serve nothing else over HTTP on the default LiveReload port" do
+      skip_if_windows "EventMachine support on Windows is limited"
       opts = serve(@standard_options)
       res = @client.get("http://#{opts["host"]}:#{opts["livereload_port"]}/")
       assert_equal(400, res.status_code)
@@ -110,6 +111,7 @@ class TestCommandsServe < JekyllUnitTest
     end
 
     should "insert the LiveReload script tags" do
+      skip_if_windows "EventMachine support on Windows is limited"
       opts = serve(@standard_options)
       content = @client.get_content(
         "http://#{opts["host"]}:#{opts["port"]}/#{opts["baseurl"]}/hello.html"
@@ -122,6 +124,7 @@ class TestCommandsServe < JekyllUnitTest
     end
 
     should "apply the max and min delay options" do
+      skip_if_windows "EventMachine support on Windows is limited"
       opts = serve(@standard_options.merge(
         "livereload_max_delay" => "1066",
         "livereload_min_delay" => "3"
